fix(dipy): float32 DWI-chain output + float32 volume loading + RAM recalibration
E9: the DWI-domain nodes computed float results and wrote them back through
the raw dcm2niix DWI's int16 header, quantizing every correction (99.9% of
voxels altered; shm_coeff rel-L2 1.08%/2.87% on subj1/subj2, WM FA up to ~0.1).
DipyDenoise, DipyMotionCorrection and DwiBiasCorrection now write float32 on
disk (affine/orientation/zooms preserved). DwiCrop stays int16 (lossless
spatial subset). The motion serial-vs-parallel oracle stays bit-for-bit.

Follow-on (user decision): load the volume nodes as float32 instead of
get_fdata's float64 default -- computing/saving float64 only to approximate at
save wastes RAM/time. DipyDenoise/DipyTensorFit/DipyCsdFit now load float32
(DwiCrop/DwiBiasCorrection/DipyTracking already did): -20 to -36% peak RAM per
node, lossless for tensor/CSD (dipy upcasts internally) and a small nlmeans
delta for denoise. DipyTissueClassifier stays float64 (dipy's HMRF Cython
kernel requires double); DipyMotionCorrection stays float64 (dipy controls the
load and float32 on the parallel path alone would break the oracle); streamline
nodes have no volume load.

RAM: measured int16->float32 is peak-neutral (nodes held float64 regardless of
on-disk dtype); float32 loading lowers peaks. Only estimator constant that
moves is DipyCsdRamEstimator.DATA_BYTES_PER_VOXEL_VOLUME (10 -> 8: float32
resident input), verified conservative at the serial and 4-worker rungs on both
subjects. Motion estimator and static _MEM_GB unchanged. Floor input does not
rise. Dipy matrix snapshot unchanged. See
docs/superpowers/2026-09-07-phase2-e9-float32-dwi.md.

+++ swane/nipype_pipeline/nodes/DipyTensorFit.py
@@ -57,7 +57,11 @@ class DipyTensorFit(BaseInterface):
         out_fa = self._gen_outfilename()
 
         in_nii = nib.load(self.inputs.in_file)
-        data = in_nii.get_fdata()
+        # Load float32 rather than get_fdata's float64 default: dipy's
+        # TensorModel upcasts to float64 internally, so FA is unchanged (~1e-7)
+        # while the input buffer's RAM is halved (user decision, 2026-09-07 --
+        # all dipy nodes load float32).
+        data = in_nii.get_fdata(dtype=np.float32)
         mask_data = nib.load(self.inputs.mask).get_fdata().astype(bool)
         bvals, bvecs = read_bvals_bvecs(self.inputs.bval, self.inputs.bvec)
         gtab = gradient_table(bvals, bvecs=bvecs)
